# backend/tests/contracts/research-runs.openapi.json
{
  "components": {
    "schemas": {
      "ComponentQuality": {
        "enum": [
          "fresh",
          "partial",
          "stale",
          "missing",
          "failed"
        ],
        "title": "ComponentQuality",
        "type": "string"
      },
      "DataQuality": {
        "additionalProperties": false,
        "properties": {
          "model": {
            "$ref": "#/components/schemas/ModelQuality"
          },
          "news": {
            "$ref": "#/components/schemas/ComponentQuality"
          },
          "overall": {
            "$ref": "#/components/schemas/OverallQuality"
          },
          "prices": {
            "$ref": "#/components/schemas/ComponentQuality"
          }
        },
        "required": [
          "overall",
          "prices",
          "news",
          "model"
        ],
        "title": "DataQuality",
        "type": "object"
      },
      "EvidenceSource": {
        "additionalProperties": false,
        "properties": {
          "content_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Hash"
          },
          "event_time": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Time"
          },
          "evidence_id": {
            "title": "Evidence Id",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "publisher": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publisher"
          },
          "retrieved_at": {
            "format": "date-time",
            "title": "Retrieved At",
            "type": "string"
          },
          "source_type": {
            "const": "news",
            "default": "news",
            "title": "Source Type",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "required": [
          "evidence_id",
          "provider",
          "title",
          "retrieved_at"
        ],
        "title": "EvidenceSource",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ModelInfo": {
        "additionalProperties": false,
        "properties": {
          "failure_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Code"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "prompt_version": {
            "title": "Prompt Version",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          }
        },
        "required": [
          "provider",
          "model",
          "prompt_version"
        ],
        "title": "ModelInfo",
        "type": "object"
      },
      "ModelQuality": {
        "enum": [
          "healthy",
          "degraded",
          "failed",
          "not_run"
        ],
        "title": "ModelQuality",
        "type": "string"
      },
      "OverallQuality": {
        "enum": [
          "sufficient",
          "degraded",
          "insufficient"
        ],
        "title": "OverallQuality",
        "type": "string"
      },
      "Reason": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "evidence_ids": {
            "items": {
              "type": "string"
            },
            "title": "Evidence Ids",
            "type": "array"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "metric_keys": {
            "items": {
              "type": "string"
            },
            "title": "Metric Keys",
            "type": "array"
          },
          "policy_version": {
            "default": "research-policy-v1",
            "title": "Policy Version",
            "type": "string"
          },
          "severity": {
            "enum": [
              "info",
              "warning",
              "blocking"
            ],
            "title": "Severity",
            "type": "string"
          },
          "threshold": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Threshold"
          }
        },
        "required": [
          "code",
          "label",
          "severity",
          "description"
        ],
        "title": "Reason",
        "type": "object"
      },
      "ResearchMetric": {
        "additionalProperties": false,
        "properties": {
          "as_of": {
            "format": "date-time",
            "title": "As Of",
            "type": "string"
          },
          "calculation_version": {
            "title": "Calculation Version",
            "type": "string"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "quality": {
            "$ref": "#/components/schemas/ComponentQuality"
          },
          "unit": {
            "enum": [
              "ratio",
              "percent",
              "usd",
              "count",
              "sessions"
            ],
            "title": "Unit",
            "type": "string"
          },
          "value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "window_sessions": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Window Sessions"
          }
        },
        "required": [
          "key",
          "label",
          "value",
          "unit",
          "window_sessions",
          "as_of",
          "calculation_version",
          "quality"
        ],
        "title": "ResearchMetric",
        "type": "object"
      },
      "ResearchRun": {
        "additionalProperties": false,
        "properties": {
          "as_of": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "data_quality": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DataQuality"
              },
              {
                "type": "null"
              }
            ]
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message_safe": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message Safe"
          },
          "metrics": {
            "items": {
              "$ref": "#/components/schemas/ResearchMetric"
            },
            "title": "Metrics",
            "type": "array"
          },
          "model_info": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModelInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "reasons": {
            "items": {
              "$ref": "#/components/schemas/Reason"
            },
            "title": "Reasons",
            "type": "array"
          },
          "research_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResearchStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/EvidenceSource"
            },
            "title": "Sources",
            "type": "array"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "symbol": {
            "pattern": "^[A-Z][A-Z0-9.]{0,9}$",
            "title": "Symbol",
            "type": "string"
          },
          "versions": {
            "$ref": "#/components/schemas/VersionInfo"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          },
          "workflow_status": {
            "$ref": "#/components/schemas/WorkflowStatus"
          }
        },
        "required": [
          "run_id",
          "symbol",
          "created_at",
          "completed_at",
          "as_of",
          "workflow_status",
          "research_status",
          "summary",
          "reasons",
          "metrics",
          "data_quality",
          "sources",
          "warnings",
          "model_info",
          "versions"
        ],
        "title": "ResearchRun",
        "type": "object"
      },
      "ResearchRunPage": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ResearchRun"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "ResearchRunPage",
        "type": "object"
      },
      "ResearchRunRequest": {
        "additionalProperties": false,
        "properties": {
          "symbol": {
            "pattern": "^[A-Z][A-Z0-9.]{0,9}$",
            "title": "Symbol",
            "type": "string"
          }
        },
        "required": [
          "symbol"
        ],
        "title": "ResearchRunRequest",
        "type": "object"
      },
      "ResearchStatus": {
        "enum": [
          "informational",
          "review",
          "insufficient_data",
          "failed"
        ],
        "title": "ResearchStatus",
        "type": "string"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VersionInfo": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "metrics": {
            "title": "Metrics",
            "type": "string"
          },
          "policy": {
            "title": "Policy",
            "type": "string"
          },
          "response_schema": {
            "title": "Response Schema",
            "type": "string"
          },
          "workflow": {
            "title": "Workflow",
            "type": "string"
          }
        },
        "required": [
          "response_schema",
          "workflow",
          "metrics",
          "policy",
          "code"
        ],
        "title": "VersionInfo",
        "type": "object"
      },
      "WorkflowStatus": {
        "enum": [
          "running",
          "succeeded",
          "failed"
        ],
        "title": "WorkflowStatus",
        "type": "string"
      }
    }
  },
  "info": {
    "title": "Vantage",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/research-runs": {
      "get": {
        "description": "List owner-scoped research runs with keyset/cursor pagination.",
        "operationId": "list_runs_api_v1_research_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "before",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Before"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchRunPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Runs",
        "tags": [
          "research-runs"
        ]
      },
      "post": {
        "description": "Create and synchronously execute a durable, bounded research run.",
        "operationId": "create_run_api_v1_research_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResearchRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchRun"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Run",
        "tags": [
          "research-runs"
        ]
      }
    },
    "/api/v1/research-runs/{run_id}": {
      "get": {
        "description": "Retrieve an owner-scoped research run by its public run_id.",
        "operationId": "get_run_api_v1_research_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchRun"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Run",
        "tags": [
          "research-runs"
        ]
      }
    }
  }
}